Add multiple fallback install variants when attempting to install cargo-sweep (#85)
On regular github actions runners installing the cargo-sweep binary for
x86_64-unknown-linux-gnu works pretty well. However, cargo-quickinstall
may not have a comparable binary that can be installed for all
architectures. This commit adds a few fallbacks in that case.

These are:
1. We try to install the musl variant of the binary
   x86_64-unknown-linux-gnu -> x86_64-unknown-linux-musl. These appear
   to be available for more architectures.
2. We install cargo-sweep via cargo-install.

It would be better to not have to use the last fallback but being a little
slow is preferable over the action failing.